feat: support for half-point client ratings & removed swagger

+++ src/main/java/org/collapseloader/atlas/domain/clients/dto/request/ClientRatingRequest.java
@@ -1,4 +1,4 @@
 package org.collapseloader.atlas.domain.clients.dto.request;
 
-public record ClientRatingRequest(short rating) {
+public record ClientRatingRequest(double rating) {
 }
